Switch 'IndicesPage' to a pure function and TypeScript.
The former addresses a Reviewbot warning.
